// src/app/enums/standartReportPeriod.enum.ts
export enum standartReportPeriod {
  CURRENT_DAY = 'current_day',
  CURRENT_WEEK = 'current_week',
  CURRENT_MONTH = 'current_month',
  PREVIOUS_DAY = 'previous_day',
  PREVIOUS_WEEK = 'previous_week',
  PREVIOUS_MONTH = 'previous_month',
  TARGET_DAY = 'target_day',
  TARGET_MONTH = 'target_month',
  TARGET_PERIOD = 'target_period'
}